backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
pydantic>=2.6
pydantic-settings>=2.2
httpx>=0.27
pyyaml>=6.0
scikit-learn>=1.4
pandas>=2.2
joblib>=1.3
python-dotenv>=1.0
numpy>=1.26
